+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Add Flatpaks to the upgrade process.
 - Change Kwallet to be disabled on the performance image.
 - Add Wine Staging again to now be used alongside Wine GE.
+- Change Arch Linux repositories to use newer upstream packages than SteamOS.
 
 ## 3.0.1